feat: docs in website!
merge the docs repo into the website at /docs.

doing this at deploy time proved tricky, so now we just build all the things together.

markdown files can be added directly to `src/pages/docs/your-topic-here.md` and add it to a `_meta.json` at the same level to set it's link title and position in the nav.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>