feat: create w3name docs (#61)
Could use thoughts on what the sidebar title for this should be. In the
old docs it was `Update data` which is clear to any audience but maybe
misleading (since content addressing, etc.). I called it `Mutate
pointers` for now to keep consistent with the other sidebar titles, but
it feels clunky.

---------

Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>