feat: update "bring your own delegations" example
1) Add missing `importDAG` and `CarReader` imports
2) Use `StoreMemory` in example to avoid annoying `Error: store cannot be used with did:key:z6xyz, stored principal and passed principal must match` error.

I started out trying to do (1) but bumped into the error in (2) after running the test code once. Customers have run into this before and we usually tell them to use `StoreMemory` - I think that's the right thing in this case anyway so this feels like a win!

I'd like to update this one more time once https://github.com/web3-storage/w3up/pull/1266 is released so we can get rid of the dependency on `@web3-storage/access`.

Fixes #63