feat: add "How to upload from CI" docs
Explain how to upload to web3.storge from CI.

The doc explains it very tersly at first so wizards may scan it quickly to get the gist, and new-comers are introduced the overall flow.

Then we give a code example for both sides for hands-on learners who just want to see the code.

Then we get to the details. This format lets the doc work for multiple different audiences without being overly verbose for folks that already have some context, or are just trying to get the thing done quickly.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>